@@ -15,7 +15,7 @@ class Convertor(typing.Generic[T]):
         raise NotImplementedError()  # pragma: no cover
 
 
-class StringConvertor(Convertor):
+class StringConvertor(Convertor[str]):
     regex = "[^/]+"
 
     def convert(self, value: str) -> str:
@@ -28,7 +28,7 @@ class StringConvertor(Convertor):
         return value
 
 
-class PathConvertor(Convertor):
+class PathConvertor(Convertor[str]):
     regex = ".*"
 
     def convert(self, value: str) -> str:
@@ -38,7 +38,7 @@ class PathConvertor(Convertor):
         return str(value)
 
 
-class IntegerConvertor(Convertor):
+class IntegerConvertor(Convertor[int]):
     regex = "[0-9]+"
 
     def convert(self, value: str) -> int:
@@ -50,7 +50,7 @@ class IntegerConvertor(Convertor):
         return str(value)
 
 
-class FloatConvertor(Convertor):
+class FloatConvertor(Convertor[float]):
     regex = r"[0-9]+(\.[0-9]+)?"
 
     def convert(self, value: str) -> float:
@@ -64,7 +64,7 @@ class FloatConvertor(Convertor):
         return ("%0.20f" % value).rstrip("0").rstrip(".")
 
 
-class UUIDConvertor(Convertor):
+class UUIDConvertor(Convertor[uuid.UUID]):
     regex = "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}"
 
     def convert(self, value: str) -> uuid.UUID:
@@ -74,7 +74,7 @@ class UUIDConvertor(Convertor):
         return str(value)
 
 
-CONVERTOR_TYPES = {
+CONVERTOR_TYPES: typing.Dict[str, Convertor[typing.Any]] = {
     "str": StringConvertor(),
     "path": PathConvertor(),
     "int": IntegerConvertor(),
@@ -83,5 +83,5 @@ CONVERTOR_TYPES = {
 }
 
 
-def register_url_convertor(key: str, convertor: Convertor) -> None:
+def register_url_convertor(key: str, convertor: Convertor[typing.Any]) -> None:
     CONVERTOR_TYPES[key] = convertor
