@@ -1,5 +1,4 @@
 import typing
-from collections.abc import Sequence
 from shlex import shlex
 from urllib.parse import SplitResult, parse_qsl, urlencode, urlsplit
 
@@ -223,7 +222,7 @@ class Secret:
         return bool(self._value)
 
 
-class CommaSeparatedStrings(Sequence):
+class CommaSeparatedStrings(typing.Sequence[str]):
     def __init__(self, value: typing.Union[str, typing.Sequence[str]]):
         if isinstance(value, str):
             splitter = shlex(value, posix=True)
@@ -269,7 +268,7 @@ class ImmutableMultiDict(typing.Mapping[_KeyType, _CovariantValueType]):
         if kwargs:
             value = (
                 ImmutableMultiDict(value).multi_items()
-                + ImmutableMultiDict(kwargs).multi_items()  # type: ignore[operator]
+                + ImmutableMultiDict(kwargs).multi_items()
             )
 
         if not value:
@@ -341,12 +340,12 @@ class MultiDict(ImmutableMultiDict[typing.Any, typing.Any]):
         self._list = [(k, v) for k, v in self._list if k != key]
         return self._dict.pop(key, default)
 
-    def popitem(self) -> typing.Tuple:
+    def popitem(self) -> typing.Tuple[typing.Any, typing.Any]:
         key, value = self._dict.popitem()
         self._list = [(k, v) for k, v in self._list if k != key]
         return key, value
 
-    def poplist(self, key: typing.Any) -> typing.List:
+    def poplist(self, key: typing.Any) -> typing.List[typing.Any]:
         values = [v for k, v in self._list if k == key]
         self.pop(key)
         return values
@@ -362,7 +361,7 @@ class MultiDict(ImmutableMultiDict[typing.Any, typing.Any]):
 
         return self[key]
 
-    def setlist(self, key: typing.Any, values: typing.List) -> None:
+    def setlist(self, key: typing.Any, values: typing.List[typing.Any]) -> None:
         if not values:
             self.pop(key, None)
         else:
@@ -378,7 +377,7 @@ class MultiDict(ImmutableMultiDict[typing.Any, typing.Any]):
         self,
         *args: typing.Union[
             "MultiDict",
-            typing.Mapping,
+            typing.Mapping[typing.Any, typing.Any],
             typing.List[typing.Tuple[typing.Any, typing.Any]],
         ],
         **kwargs: typing.Any,
@@ -397,8 +396,8 @@ class QueryParams(ImmutableMultiDict[str, str]):
     def __init__(
         self,
         *args: typing.Union[
-            "ImmutableMultiDict",
-            typing.Mapping,
+            "ImmutableMultiDict[typing.Any, typing.Any]",
+            typing.Mapping[typing.Any, typing.Any],
             typing.List[typing.Tuple[typing.Any, typing.Any]],
             str,
             bytes,
@@ -482,6 +481,14 @@ class UploadFile:
             self.file.close()
         else:
             await run_in_threadpool(self.file.close)
+
+    def __repr__(self) -> str:
+        return (
+            f"{self.__class__.__name__}("
+            f"filename={self.filename!r}, "
+            f"size={self.size!r}, "
+            f"headers={self.headers!r})"
+        )
 
 
 class FormData(ImmutableMultiDict[str, typing.Union[UploadFile, str]]):
