@@ -10,13 +10,16 @@ class HTTPException(Exception):
         self,
         status_code: int,
         detail: typing.Optional[str] = None,
-        headers: typing.Optional[dict] = None,
+        headers: typing.Optional[typing.Dict[str, str]] = None,
     ) -> None:
         if detail is None:
             detail = http.HTTPStatus(status_code).phrase
         self.status_code = status_code
         self.detail = detail
         self.headers = headers
+
+    def __str__(self) -> str:
+        return f"{self.status_code}: {self.detail}"
 
     def __repr__(self) -> str:
         class_name = self.__class__.__name__
@@ -27,6 +30,9 @@ class WebSocketException(Exception):
     def __init__(self, code: int, reason: typing.Optional[str] = None) -> None:
         self.code = code
         self.reason = reason or ""
+
+    def __str__(self) -> str:
+        return f"{self.code}: {self.reason}"
 
     def __repr__(self) -> str:
         class_name = self.__class__.__name__
