@@ -24,7 +24,7 @@ class AuthenticationMiddleware:
         self.backend = backend
         self.on_error: typing.Callable[
             [HTTPConnection, AuthenticationError], Response
-        ] = (on_error if on_error is not None else self.default_on_error)
+        ] = on_error if on_error is not None else self.default_on_error
 
     async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
         if scope["type"] not in ["http", "websocket"]:
