@@ -137,7 +137,9 @@ class ServerErrorMiddleware:
     def __init__(
         self,
         app: ASGIApp,
-        handler: typing.Optional[typing.Callable] = None,
+        handler: typing.Optional[
+            typing.Callable[[Request, Exception], typing.Any]
+        ] = None,
         debug: bool = False,
     ) -> None:
         self.app = app
@@ -199,7 +201,10 @@ class ServerErrorMiddleware:
     def generate_frame_html(self, frame: inspect.FrameInfo, is_collapsed: bool) -> str:
         code_context = "".join(
             self.format_line(
-                index, line, frame.lineno, frame.index  # type: ignore[arg-type]
+                index,
+                line,
+                frame.lineno,
+                frame.index,  # type: ignore[arg-type]
             )
             for index, line in enumerate(frame.code_context or [])
         )
