@@ -1,5 +1,4 @@
 import json
-import sys
 import typing
 from base64 import b64decode, b64encode
 
@@ -10,11 +9,6 @@ from starlette.datastructures import MutableHeaders, Secret
 from starlette.requests import HTTPConnection
 from starlette.types import ASGIApp, Message, Receive, Scope, Send
 
-if sys.version_info >= (3, 8):  # pragma: no cover
-    from typing import Literal
-else:  # pragma: no cover
-    from typing_extensions import Literal
-
 
 class SessionMiddleware:
     def __init__(
@@ -24,8 +18,9 @@ class SessionMiddleware:
         session_cookie: str = "session",
         max_age: typing.Optional[int] = 14 * 24 * 60 * 60,  # 14 days, in seconds
         path: str = "/",
-        same_site: Literal["lax", "strict", "none"] = "lax",
+        same_site: typing.Literal["lax", "strict", "none"] = "lax",
         https_only: bool = False,
+        domain: typing.Optional[str] = None,
     ) -> None:
         self.app = app
         self.signer = itsdangerous.TimestampSigner(str(secret_key))
@@ -35,6 +30,8 @@ class SessionMiddleware:
         self.security_flags = "httponly; samesite=" + same_site
         if https_only:  # Secure flag can be used with HTTPS only
             self.security_flags += "; secure"
+        if domain is not None:
+            self.security_flags += f"; domain={domain}"
 
     async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
         if scope["type"] not in ("http", "websocket"):  # pragma: no cover
